voice-bridge/pi-client/requirements.txt
requests>=2.28.0
RPi.GPIO>=0.7.0
pyaudio>=0.2.13
